hash.c: replace some unsigned long with unsigned int
	* hash.c (higher_prime_number): Use uint32_t param, return value,
	tables and variables.
	(bfd_default_hash_table_size): Make it an unsigned int.
	(bfd_hash_set_default_size): Use unsigned int param and return.
	* bfd-in.h (bfd_hash_set_default_size): Update prototype.
	* bfd-in2.h: Regenerate.

+++ bfd/bfd-in2.h
@@ -340,7 +340,7 @@ extern void bfd_hash_traverse
 /* Allows the default size of a hash table to be configured. New hash
    tables allocated using bfd_hash_table_init will be created with
    this size.  */
-extern unsigned long bfd_hash_set_default_size (unsigned long);
+extern unsigned int bfd_hash_set_default_size (unsigned int);
 
 /* This structure is used to keep track of stabs in sections
    information while linking.  */
